I fixed a bug where the Manage Data button was in front of the Setup Guide control at startup.
I also updated the project template installer to read 'DataTier.Net Project Templates', instead of DataTier.Net or DataTierNetTemplateInstaller.

And finally I added the version number to display on the Main Form at startup. I realized today I never built an About form, so displaying the version in the title at least gives people a way to know if they have the latest version or not.