I updated the SetupControl to support encrypting the app.config values.
I also fixed a bug that had been in for a while where the custom Encryption Key was not working.

This is one of the good things of not many people using DataTier.Net (yet) is bugs can be found.

I also decided to encrypt the EncryptionKey in the app.config if someone elects to use a custom encryuption key.

I now need to update the Templates to support an encrypted EncryptionKey.

Now that I added a SetupControl, all the documentation has to be updated with new screen shots, but it should be much simpler to explain now that it is down to 3 steps.