Version 1.2.5:
I added some code to fix a table that had been deleted was still saved in the DTNTables for a project.

I added the code earlier to load the fields for the existing table, but I wasn't checking for null, so if a table didn't have any fields it would throw an error.