I just realized that I had code referring to a new Admin object, but I didn't have any place where I installed the Admin object if it doesn't exist.
I added a check to the Check For Updates method that if the Admin doesn't exist, it is created.

Also in the Init method where I test if CheckForUpdate is true, then I call CheckForUpdate, if the Admin object was not found I call CheckForUpdate so the Admin object can be created.