My first version of a Custom Where Clause editor didn't work.
I changed the way the BaseMethodName is set, and it has passes 1 test so far, but better than yesterday where it didn't work.

I also fixed a bug where if the single field parameter is the primary key, the code was attempting to set the value, which is read only, you have to call UpdateIdentity instead.