I added a new field to the table DTNTable.Excluded.
This field is used to determine if the IncludeProjectFiles wizard needs to launch for a Dot Net 5 project, if it was previously excluded. The real cause is this file does not need to be removed the project for Dot Net 5 projects, I may look at that next. I thought of that after I fixed it this way.